add multi select for image in rooms and add create booking

+++ database/migrations/2024_06_28_151206_create_bookings_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('homestays_id');
             $table->foreignId('rooms_id');
             $table->integer('day');
-            $table->double('total_price');
+            $table->decimal('total_price', 10, 2);
             $table->enum('status',['process','sucesss','abort']);
             $table->timestamps();
         });
